feat: make cloud movement based on world time with options

+++ common/src/main/java/com/qendolin/betterclouds/clouds/ChunkedGenerator.java
@@ -144,7 +144,7 @@ public class ChunkedGenerator implements AutoCloseable {
         clear();
     }
 
-    public synchronized void update(Vector3d camera, int ticks, float tickDelta, Config options, float cloudiness) {
+    public synchronized void update(Vector3d camera, long ticks, float tickDelta, Config options, float cloudiness) {
         originX = RandomPath.getPathX(Math.abs(ticks + tickDelta), Math.abs(options.travelSpeed));
         originZ = RandomPath.getPathZ(Math.abs(ticks + tickDelta), Math.abs(options.travelSpeed));
